+++ testing/helpers.py
@@ -24,10 +24,13 @@ def parse_source(src: str) -> Tuple[ast.Module, List[str]]:
     return tree, lines
 
 
-def check_source(src: str) -> Generator[Error, None, None]:
+def check_source(src: str, suppress_none_returns: bool = False) -> Generator[Error, None, None]:
     """Helper for generating linting errors from the provided source code."""
     _, lines = parse_source(src)
     checker_instance = TypeHintChecker(None, lines)
+
+    # Manually set None return error suppression, as the test suite bypasses flake8's config parser
+    checker_instance.suppress_none_returning = suppress_none_returns
 
     return checker_instance.run()